# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Character Chat Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface Character {
  id: string;
  slug: string;
  name: string;
  description: string;
  /** @nullable */
  bio?: string | null;
  tags: string[];
  category: string;
  isOnline: boolean;
  /** @nullable */
  avatarUrl: string | null;
  personalityTraits: string[];
  conversationStarters: string[];
}

export type MessageRole = (typeof MessageRole)[keyof typeof MessageRole];

export const MessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface Message {
  id: string;
  characterSlug: string;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export interface SendMessageRequest {
  /**
   * @minLength 1
   * @maxLength 2000
   */
  content: string;
}

export interface AccountInfo {
  user: AuthUser;
  favorites: string[];
  totalChats: number;
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  /**
   * Relative path to redirect to after login (must start with `/`). Defaults to `/`.
   */
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
  iss?: string;
};

export type ListCharactersParams = {
  search?: string;
  tag?: string;
  category?: string;
};

export type ListCharacters200 = {
  characters: Character[];
};

export type GetChatHistory200 = {
  messages: Message[];
};

export type GetFavorites200 = {
  favorites: string[];
};