# lib/api-zod/src/generated/types/accountInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Character Chat Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUser } from "./authUser";

export interface AccountInfo {
  user: AuthUser;
  favorites: string[];
  totalChats: number;
}