# lib/api-zod/src/generated/types/character.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Character Chat Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Character {
  id: string;
  slug: string;
  name: string;
  description: string;
  /** @nullable */
  bio?: string | null;
  tags: string[];
  category: string;
  isOnline: boolean;
  /** @nullable */
  avatarUrl: string | null;
  personalityTraits: string[];
  conversationStarters: string[];
}